[REF] mail,*:tests: clean test files
