+++ addons/mail/static/tests/new/web/fields/emojis_text_field_tests.js
@@ -12,34 +12,35 @@ import {
 
 let target = undefined;
 
-addFakeModel("fields.char.emojis.user", { foo: { type: "char", onChange: "1" } });
+addFakeModel("fields.text.emojis.user", {
+    foo: { type: "char", onChange: "1" },
+});
 const views = {
-    "fields.char.emojis.user,false,form": `
+    "fields.text.emojis.user,false,form": `
         <form>
-            <field name="foo" widget="char_emojis"/>
+            <field name="foo" widget="text_emojis"/>
         </form>
     `,
 };
 async function openTestView(readonly = false) {
     const pyEnv = await startServer();
-    const recordId = pyEnv["fields.char.emojis.user"].create({
+    const recordId = pyEnv["fields.text.emojis.user"].create({
         display_name: "test record",
         foo: "test",
     });
-    const startServerArgs = { serverData: { views } };
     const openViewArgs = {
         res_id: recordId,
-        res_model: "fields.char.emojis.user",
+        res_model: "fields.text.emojis.user",
         views: [[false, "form"]],
     };
     if (readonly) {
         openViewArgs.context = { form_view_initial_mode: "readonly" };
     }
-    const { openView } = await start(startServerArgs);
+    const { openView } = await start({ serverData: { views } });
     await openView(openViewArgs);
 }
 
-QUnit.module("Field char emojis", {
+QUnit.module("Field text emojis", {
     beforeEach() {
         target = getFixture();
     },
@@ -47,19 +48,17 @@ QUnit.module("Field char emojis", {
 
 QUnit.test("emojis button is not shown in readonly mode", async (assert) => {
     await openTestView(true);
-    await testEmojiButtonHidden(assert, target, ".o_field_char_emojis");
+    await testEmojiButtonHidden(assert, target, ".o_field_text_emojis");
 });
 
 QUnit.test("emojis button is shown in edit mode", async (assert) => {
     await openTestView();
-    await testEmojiButtonVisible(assert, target, ".o_field_char_emojis");
+    await testEmojiButtonVisible(assert, target, ".o_field_text_emojis");
 });
 
 QUnit.test("emojis button works", async (assert) => {
     await openTestView();
-
-    const input = target.querySelector(".o_field_char_emojis input[type='text']");
-    const emojiButton = target.querySelector(".o_field_char_emojis button");
-
+    const input = target.querySelector(".o_field_text_emojis textarea");
+    const emojiButton = target.querySelector(".o_field_text_emojis button");
     await testEmojiButton(assert, target, input, emojiButton);
 });
