+++ addons/sms/__manifest__.py
@@ -51,7 +51,6 @@ The service is provided by the In App Purchase Odoo platform.
             'sms/static/src/components/*/*',
         ],
         'web.qunit_suite_tests': [
-            'sms/static/tests/qunit_suite_tests/**/*.js',
             'sms/static/tests/new/**/*',
         ],
     },
